# docs/openapi/v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Eldercare Edge Enrollment and Topology API",
    "version": "1.0.0",
    "description": "Canonical frozen v1 contract for edge credential administration, installation enrollment, edge-authored topology reconciliation, machine ingest/config access, and audited clip download. Server-derived facility, installation, and enrollment-generation identity is authoritative on every edge-bearer request."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API server"
    }
  ],
...
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "operationId": "createMachineEvent",
        "tags": [
          "Machine API"
        ],
        "summary": "Record an ML event",
        "description": "The eft_v1 principal supplies facility, installation, and generation. Camera ownership is resolved server-side. Facility identity fields are not part of the v1 body. Previously deployed shared-token clients remain a separately metered compatibility path and do not gain eft_v1 authority.",
        "security": [
          {
            "edgeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEventResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidSchema"
          },
          "401": {
            "$ref": "#/components/responses/EdgeUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/EdgeForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/AuthNotConfigured"
          }
        }
      }
    },
    "/api/v1/events/heartbeat": {
      "post": {
        "operationId": "recordMachineHeartbeat",
        "tags": [
          "Machine API"
        ],
        "summary": "Record a camera heartbeat",
        "description": "Marks the server-resolved camera online under the authenticated edge principal. No facility header or body field is accepted for eft_v1.",
        "security": [
          {
            "edgeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidSchema"
          },
          "401": {
            "$ref": "#/components/responses/EdgeUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/EdgeForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/AuthNotConfigured"
          }
        }
      }
    },
    "/api/v1/events/{eventId}/snapshot": {
      "put": {
        "operationId": "uploadMachineEventSnapshot",
        "tags": [
          "Machine API"
        ],
        "summary": "Upload immutable event snapshot bytes",
        "description": "Resolves event ownership under the authenticated edge principal and stores bytes under a server-derived key. Client-supplied snapshot keys and facility identity headers are rejected.",
        "security": [
          {
            "edgeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "contentEncoding": "binary"
              }
            },
            "image/png": {
              "schema": {
                "type": "string",
                "contentEncoding": "binary"
              }
            },
            "image/webp": {
              "schema": {
                "type": "string",
                "contentEncoding": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exact bytes already stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotUploadResponse"
                }
              }
            }
          },
          "201": {
            "description": "Immutable snapshot created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidSchema"
          },
          "401": {
            "$ref": "#/components/responses/EdgeUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/EdgeForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/AuthNotConfigured"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "app_session",
        "description": "HttpOnly product session. Operation metadata further restricts the required role or capability."
      },
      "edgeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "eft_v1.{tokenId}.{one-time-value}",
        "description": "Installation- and enrollment-generation-bound v1 edge principal. A legacy shared token is not authorized by this security scheme."
      }
    },
...
    "headers": {
      "RequestId": {
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
    },
    "responses": {
      "InvalidSchema": {
        "description": "Malformed or unsupported request. Code is INVALID_SCHEMA and retryable is false.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "EdgeUnauthorized": {
        "description": "Generic EDGE_CREDENTIAL_REQUIRED or EDGE_CREDENTIAL_INVALID response with equivalent timing and no record disclosure. Retryable is false.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "EdgeForbidden": {
        "description": "Generic EDGE_CREDENTIAL_INACTIVE or FACILITY_BINDING_MISMATCH response. This includes revoked, expired, stale-generation, and foreign-scope principals where disclosure would expose state. Retryable is false.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "State or idempotency conflict. Permitted codes are INSTALLATION_CONFLICT, IDEMPOTENCY_CONFLICT, CLIENT_REVISION_OUT_OF_SEQUENCE, STALE_SERVER_REVISION, STALE_ENROLLMENT_GENERATION, TOPOLOGY_CONFLICT, TOPOLOGY_TRANSFER_CONFLICT, LEGACY_MAPPING_REQUIRED, or CONFIRMATION_STALE. Retryable is false; refresh state or take explicit operator action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "AuthNotConfigured": {
        "description": "EDGE_AUTH_NOT_CONFIGURED. The host token verifier is unavailable or has no pepper; retryable is true.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Rfc3339MillisUtc": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
        "description": "RFC 3339 UTC timestamp with exactly millisecond precision.",
        "example": "2026-01-01T00:00:00.000Z"
      },
...
      "EdgeErrorCode": {
        "type": "string",
        "enum": [
          "INVALID_SCHEMA",
          "INVALID_TOPOLOGY",
          "EDGE_CREDENTIAL_REQUIRED",
          "EDGE_CREDENTIAL_INVALID",
          "EDGE_CREDENTIAL_INACTIVE",
          "FACILITY_BINDING_MISMATCH",
          "INSTALLATION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "CLIENT_REVISION_OUT_OF_SEQUENCE",
          "STALE_SERVER_REVISION",
          "STALE_ENROLLMENT_GENERATION",
          "TOPOLOGY_CONFLICT",
          "TOPOLOGY_TRANSFER_CONFLICT",
          "LEGACY_MAPPING_REQUIRED",
          "CONFIRMATION_STALE",
          "CONFIRMATION_EXPIRED",
          "ENROLLMENT_RATE_LIMITED",
          "EDGE_AUTH_NOT_CONFIGURED"
        ]
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schemaVersion",
          "error"
        ],
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "$ref": "#/components/schemas/EdgeError"
          }
        }
      },
      "EdgeError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/EdgeErrorCode"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256,
            "description": "Generic, non-disclosing message."
          },
          "retryable": {
            "type": "boolean",
            "description": "True only for ENROLLMENT_RATE_LIMITED and EDGE_AUTH_NOT_CONFIGURED server responses. Client-observed network/timeouts are separately retryable transport failures."
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "allOf": [
          {
            "if": {
              "required": [
                "code"
              ],
              "properties": {
                "code": {
                  "enum": [
                    "ENROLLMENT_RATE_LIMITED",
                    "EDGE_AUTH_NOT_CONFIGURED"
                  ]
                }
              }
            },
            "then": {
              "properties": {
                "retryable": {
                  "const": true
                }
              }
            },
            "else": {
              "properties": {
                "retryable": {
                  "const": false
                }
              }
            }
          }
        ]
      },
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "example": "a5ff4ed1-7e63-4a4f-9ef0-42e807d74a64"
      },
...
      "RecordEventRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "camera_id",
          "type",
          "detected_at"
        ],
        "properties": {
          "camera_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "detected_at": {
            "$ref": "#/components/schemas/Rfc3339MillisUtc"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "config_version": {
            "type": "number",
            "minimum": 0
          },
          "model_version": {
            "type": "string",
            "maxLength": 128
          },
          "detector_version": {
            "type": "string",
            "maxLength": 128
          },
          "operating_threshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "snapshot_key": {
            "oneOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ]
          },
          "clip_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "clock_source": {
            "type": "string",
            "maxLength": 64
          },
          "edge_event_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "description": "No facility identifier, connection URL, camera authentication material, or secret-shaped field is accepted."
      },
      "RecordEventResponse": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "status"
            ],
            "properties": {
              "id": {
                "$ref": "#/components/schemas/Uuid"
              },
              "status": {
                "type": "string",
                "enum": [
                  "created",
                  "duplicate"
                ]
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "event_id",
              "edge_event_id",
              "status"
            ],
            "properties": {
              "id": {
                "$ref": "#/components/schemas/Uuid"
              },
              "event_id": {
                "$ref": "#/components/schemas/Uuid"
              },
              "edge_event_id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "const": "accepted"
              }
            }
          }
        ]
      },
      "HeartbeatRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "camera_id"
        ],
        "properties": {
          "camera_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
      "HeartbeatResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "SnapshotUploadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "snapshotKey"
        ],
        "properties": {
          "snapshotKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "readOnly": true,
            "description": "Opaque server-derived storage key."
          }
        }
      },
...
    }
  }
}